Split formatters into separate files
