+++ lib/stackprof/formatter/debug.rb
@@ -1,0 +1,9 @@
+module StackProf
+  module Formatter
+    module Debug
+      def print_debug
+        pp @data
+      end
+    end
+  end
+end
