+++ lib/stackprof/report.rb
@@ -1,8 +1,28 @@
 require 'pp'
 require 'digest/md5'
 
+require_relative 'formatter/callgrind'
+require_relative 'formatter/d3_flamegraph'
+require_relative 'formatter/debug'
+require_relative 'formatter/dump'
+require_relative 'formatter/flamegraph'
+require_relative 'formatter/graphviz'
+require_relative 'formatter/json'
+require_relative 'formatter/stackcollapse'
+require_relative 'formatter/text'
+
 module StackProf
   class Report
+    include Formatter::Callgrind
+    include Formatter::D3Flamegraph
+    include Formatter::Debug
+    include Formatter::Dump
+    include Formatter::Flamegraph
+    include Formatter::Graphviz
+    include Formatter::Json
+    include Formatter::Stackcollapse
+    include Formatter::Text
+
     def initialize(data)
       @data = data
     end
@@ -58,464 +78,6 @@ module StackProf
       return a+b if a.is_a? Integer
       return [ a[0], a[1]+b ] if b.is_a? Integer
       [ a[0]+b[0], a[1]+b[1] ]
-    end
-
-    def print_debug
-      pp @data
-    end
-
-    def print_dump(f=STDOUT)
-      f.puts Marshal.dump(@data.reject{|k,v| k == :files })
-    end
-
-    def print_json(f=STDOUT)
-      require "json"
-      f.puts JSON.generate(@data, max_nesting: false)
-    end
-
-    def print_stackcollapse
-      raise "profile does not include raw samples (add `raw: true` to collecting StackProf.run)" unless raw = data[:raw]
-
-      while len = raw.shift
-        frames = raw.slice!(0, len)
-        weight = raw.shift
-
-        print frames.map{ |a| data[:frames][a][:name] }.join(';')
-        puts " #{weight}"
-      end
-    end
-
-    def print_timeline_flamegraph(f=STDOUT, skip_common=true)
-      print_flamegraph(f, skip_common, false)
-    end
-
-    def print_alphabetical_flamegraph(f=STDOUT, skip_common=true)
-      print_flamegraph(f, skip_common, true)
-    end
-
-    def print_flamegraph(f, skip_common, alphabetical=false)
-      raise "profile does not include raw samples (add `raw: true` to collecting StackProf.run)" unless raw = data[:raw]
-
-      stacks = []
-      max_x = 0
-      max_y = 0
-      while len = raw.shift
-        max_y = len if len > max_y
-        stack = raw.slice!(0, len+1)
-        stacks << stack
-        max_x += stack.last
-      end
-
-      stacks.sort! if alphabetical
-
-      f.puts 'flamegraph(['
-      max_y.times do |y|
-        row_prev = nil
-        row_width = 0
-        x = 0
-
-        stacks.each do |stack|
-          weight = stack.last
-          cell = stack[y] unless y == stack.length-1
-
-          if cell.nil?
-            if row_prev
-              flamegraph_row(f, x - row_width, y, row_width, row_prev)
-            end
-
-            row_prev = nil
-            x += weight
-            next
-          end
-
-          if row_prev.nil?        # start new row with this cell
-            row_width = weight
-            row_prev = cell
-            x += weight
-
-          elsif row_prev == cell  # grow current row along x-axis
-            row_width += weight
-            x += weight
-
-          else                    # end current row and start new row
-            flamegraph_row(f, x - row_width, y, row_width, row_prev)
-            x += weight
-            row_prev = cell
-            row_width = weight
-          end
-
-          row_prev = cell
-        end
-
-        if row_prev
-          next if skip_common && row_width == max_x
-
-          flamegraph_row(f, x - row_width, y, row_width, row_prev)
-        end
-      end
-      f.puts '])'
-    end
-
-    def flamegraph_row(f, x, y, weight, addr)
-      frame = frames[addr]
-      f.print ',' if @rows_started
-      @rows_started = true
-      f.puts %{{"x":#{x},"y":#{y},"width":#{weight},"frame_id":#{addr},"frame":#{frame[:name].dump},"file":#{frame[:file].dump}}}
-    end
-
-    def convert_to_d3_flame_graph_format(name, stacks, depth)
-      weight = 0
-      children = []
-      stacks.chunk do |stack|
-        if depth == stack.length - 1
-          :leaf
-        else
-          stack[depth]
-        end
-      end.each do |val, child_stacks|
-        if val == :leaf
-          child_stacks.each do |stack|
-            weight += stack.last
-          end
-        else
-          frame = frames[val]
-          child_name = "#{ frame[:name] } : #{ frame[:file] }"
-          child_data = convert_to_d3_flame_graph_format(child_name, child_stacks, depth + 1)
-          weight += child_data["value"]
-          children << child_data
-        end
-      end
-
-      {
-        "name" => name,
-        "value" => weight,
-        "children" => children,
-      }
-    end
-
-    def print_d3_flamegraph(f=STDOUT, skip_common=true)
-      raise "profile does not include raw samples (add `raw: true` to collecting StackProf.run)" unless raw = data[:raw]
-
-      stacks = []
-      max_x = 0
-      max_y = 0
-      while len = raw.shift
-        max_y = len if len > max_y
-        stack = raw.slice!(0, len+1)
-        stacks << stack
-        max_x += stack.last
-      end
-
-      # d3-flame-grpah supports only alphabetical flamegraph
-      stacks.sort!
-
-      require "json"
-      json = JSON.generate(convert_to_d3_flame_graph_format("<root>", stacks, 0), max_nesting: false)
-
-      # This html code is almost copied from d3-flame-graph sample code.
-      # (Apache License 2.0)
-      # https://github.com/spiermar/d3-flame-graph/blob/gh-pages/index.html
-
-      f.print <<-END
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-    <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/gh/spiermar/d3-flame-graph@2.0.3/dist/d3-flamegraph.css">
-
-    <style>
-
-    /* Space out content a bit */
-    body {
-      padding-top: 20px;
-      padding-bottom: 20px;
-    }
-
-    /* Custom page header */
-    .header {
-      padding-bottom: 20px;
-      padding-right: 15px;
-      padding-left: 15px;
-      border-bottom: 1px solid #e5e5e5;
-    }
-
-    /* Make the masthead heading the same height as the navigation */
-    .header h3 {
-      margin-top: 0;
-      margin-bottom: 0;
-      line-height: 40px;
-    }
-
-    /* Customize container */
-    .container {
-      max-width: 990px;
-    }
-
-    address {
-      text-align: right;
-    }
-    </style>
-
-    <title>stackprof (mode: #{ data[:mode] })</title>
-
-    <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
-    <!--[if lt IE 9]>
-      <script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
-      <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
-    <![endif]-->
-  </head>
-  <body>
-    <div class="container">
-      <div class="header clearfix">
-        <nav>
-          <div class="pull-right">
-            <form class="form-inline" id="form">
-              <a class="btn" href="javascript: resetZoom();">Reset zoom</a>
-              <a class="btn" href="javascript: clear();">Clear</a>
-              <div class="form-group">
-                <input type="text" class="form-control" id="term">
-              </div>
-              <a class="btn btn-primary" href="javascript: search();">Search</a>
-            </form>
-          </div>
-        </nav>
-        <h3 class="text-muted">stackprof (mode: #{ data[:mode] })</h3>
-      </div>
-      <div id="chart">
-      </div>
-      <address>
-        powered by <a href="https://github.com/spiermar/d3-flame-graph">d3-flame-graph</a>
-      </address>
-      <hr>
-      <div id="details">
-      </div>
-    </div>
-
-    <!-- D3.js -->
-    <script src="https://d3js.org/d3.v4.min.js" charset="utf-8"></script>
-
-    <!-- d3-tip -->
-    <script type="text/javascript" src=https://cdnjs.cloudflare.com/ajax/libs/d3-tip/0.9.1/d3-tip.min.js></script>
-
-    <!-- d3-flamegraph -->
-    <script type="text/javascript" src="https://cdn.jsdelivr.net/gh/spiermar/d3-flame-graph@2.0.3/dist/d3-flamegraph.min.js"></script>
-
-    <script type="text/javascript">
-    var flameGraph = d3.flamegraph()
-      .width(960)
-      .cellHeight(18)
-      .transitionDuration(750)
-      .minFrameSize(5)
-      .transitionEase(d3.easeCubic)
-      .sort(true)
-      //Example to sort in reverse order
-      //.sort(function(a,b){ return d3.descending(a.name, b.name);})
-      .title("")
-      .onClick(onClick)
-      .differential(false)
-      .selfValue(false);
-
-
-    // Example on how to use custom tooltips using d3-tip.
-    // var tip = d3.tip()
-    //   .direction("s")
-    //   .offset([8, 0])
-    //   .attr('class', 'd3-flame-graph-tip')
-    //   .html(function(d) { return "name: " + d.data.name + ", value: " + d.data.value; });
-
-    // flameGraph.tooltip(tip);
-
-    var details = document.getElementById("details");
-    flameGraph.setDetailsElement(details);
-
-    // Example on how to use custom labels
-    // var label = function(d) {
-    //  return "name: " + d.name + ", value: " + d.value;
-    // }
-    // flameGraph.label(label);
-
-    // Example of how to set fixed chart height
-    // flameGraph.height(540);
-
-    d3.select("#chart")
-        .datum(#{ json })
-        .call(flameGraph);
-
-    document.getElementById("form").addEventListener("submit", function(event){
-      event.preventDefault();
-      search();
-    });
-
-    function search() {
-      var term = document.getElementById("term").value;
-      flameGraph.search(term);
-    }
-
-    function clear() {
-      document.getElementById('term').value = '';
-      flameGraph.clear();
-    }
-
-    function resetZoom() {
-      flameGraph.resetZoom();
-    }
-
-    function onClick(d) {
-      console.info("Clicked on " + d.data.name);
-    }
-    </script>
-  </body>
-</html>
-      END
-    end
-
-    def print_graphviz(options = {}, f = STDOUT)
-      if filter = options[:filter]
-        mark_stack = []
-        list = frames(true)
-        list.each{ |addr, frame| mark_stack << addr if frame[:name] =~ filter }
-        while addr = mark_stack.pop
-          frame = list[addr]
-          unless frame[:marked]
-            mark_stack += frame[:edges].map{ |addr, weight| addr if list[addr][:total_samples] <= weight*1.2 }.compact if frame[:edges]
-            frame[:marked] = true
-          end
-        end
-        list = list.select{ |addr, frame| frame[:marked] }
-        list.each{ |addr, frame| frame[:edges] && frame[:edges].delete_if{ |k,v| list[k].nil? } }
-        list
-      else
-        list = frames(true)
-      end
-
-
-      limit = options[:limit]
-      fraction = options[:node_fraction]
-
-      included_nodes = {}
-      node_minimum = fraction ? (fraction * overall_samples).ceil : 0
-
-      f.puts "digraph profile {"
-      f.puts "Legend [shape=box,fontsize=24,shape=plaintext,label=\""
-      f.print "Total samples: #{overall_samples}\\l"
-      f.print "Showing top #{limit} nodes\\l" if limit
-      f.print "Dropped nodes with < #{node_minimum} samples\\l" if fraction
-      f.puts "\"];"
-
-      list.each_with_index do |(frame, info), index|
-        call, total = info.values_at(:samples, :total_samples)
-        break if total < node_minimum || (limit && index >= limit)
-
-        sample = ''
-        sample << "#{call} (%2.1f%%)\\rof " % (call*100.0/overall_samples) if call < total
-        sample << "#{total} (%2.1f%%)\\r" % (total*100.0/overall_samples)
-        fontsize = (1.0 * call / max_samples) * 28 + 10
-        size = (1.0 * total / overall_samples) * 2.0 + 0.5
-
-        f.puts "  \"#{frame}\" [size=#{size}] [fontsize=#{fontsize}] [penwidth=\"#{size}\"] [shape=box] [label=\"#{info[:name]}\\n#{sample}\"];"
-        included_nodes[frame] = true
-      end
-
-      list.each do |frame, info|
-        next unless included_nodes[frame]
-
-        if edges = info[:edges]
-          edges.each do |edge, weight|
-            next unless included_nodes[edge]
-
-            size = (1.0 * weight / overall_samples) * 2.0 + 0.5
-            f.puts "  \"#{frame}\" -> \"#{edge}\" [label=\"#{weight}\"] [weight=\"#{weight}\"] [penwidth=\"#{size}\"];"
-          end
-        end
-      end
-      f.puts "}"
-    end
-
-    def print_text(sort_by_total=false, limit=nil, select_files= nil, reject_files=nil, select_names=nil, reject_names=nil, f = STDOUT)
-      f.puts "=================================="
-      f.printf "  Mode: #{modeline}\n"
-      f.printf "  Samples: #{@data[:samples]} (%.2f%% miss rate)\n", 100.0*@data[:missed_samples]/(@data[:missed_samples]+@data[:samples])
-      f.printf "  GC: #{@data[:gc_samples]} (%.2f%%)\n", 100.0*@data[:gc_samples]/@data[:samples]
-      f.puts "=================================="
-      f.printf "% 10s    (pct)  % 10s    (pct)     FRAME\n" % ["TOTAL", "SAMPLES"]
-      list = frames(sort_by_total)
-      list.select!{|_, info| select_files.any?{|path| info[:file].start_with?(path)}} if select_files
-      list.select!{|_, info| select_names.any?{|reg| info[:name] =~ reg}} if select_names
-      list.reject!{|_, info| reject_files.any?{|path| info[:file].start_with?(path)}} if reject_files
-      list.reject!{|_, info| reject_names.any?{|reg| info[:name] =~ reg}} if reject_names
-      list = list.first(limit) if limit
-      list.each do |frame, info|
-        call, total = info.values_at(:samples, :total_samples)
-        f.printf "% 10d % 8s  % 10d % 8s     %s\n", total, "(%2.1f%%)" % (total*100.0/overall_samples), call, "(%2.1f%%)" % (call*100.0/overall_samples), info[:name]
-      end
-    end
-
-    def print_callgrind(f = STDOUT)
-      f.puts "version: 1"
-      f.puts "creator: stackprof"
-      f.puts "pid: 0"
-      f.puts "cmd: ruby"
-      f.puts "part: 1"
-      f.puts "desc: mode: #{modeline}"
-      f.puts "desc: missed: #{@data[:missed_samples]})"
-      f.puts "positions: line"
-      f.puts "events: Instructions"
-      f.puts "summary: #{@data[:samples]}"
-
-      list = frames
-      list.each do |addr, frame|
-        f.puts "fl=#{frame[:file]}"
-        f.puts "fn=#{frame[:name]}"
-        frame[:lines].each do |line, weight|
-          f.puts "#{line} #{weight.is_a?(Array) ? weight[1] : weight}"
-        end if frame[:lines]
-        frame[:edges].each do |edge, weight|
-          oframe = list[edge]
-          f.puts "cfl=#{oframe[:file]}" unless oframe[:file] == frame[:file]
-          f.puts "cfn=#{oframe[:name]}"
-          f.puts "calls=#{weight} #{frame[:line] || 0}\n#{oframe[:line] || 0} #{weight}"
-        end if frame[:edges]
-        f.puts
-      end
-
-      f.puts "totals: #{@data[:samples]}"
-    end
-
-    def print_method(name, f = STDOUT)
-      name = /#{name}/ unless Regexp === name
-      frames.each do |frame, info|
-        next unless info[:name] =~ name
-        file, line = info.values_at(:file, :line)
-        line ||= 1
-
-        lines = info[:lines]
-        maxline = lines ? lines.keys.max : line + 5
-        f.printf "%s (%s:%d)\n", info[:name], file, line
-        f.printf "  samples: % 5d self (%2.1f%%)  /  % 5d total (%2.1f%%)\n", info[:samples], 100.0*info[:samples]/overall_samples, info[:total_samples], 100.0*info[:total_samples]/overall_samples
-
-        if (callers = callers_for(frame)).any?
-          f.puts "  callers:"
-          callers = callers.sort_by(&:last).reverse
-          callers.each do |name, weight|
-            f.printf "   % 5d  (% 8s)  %s\n", weight, "%3.1f%%" % (100.0*weight/info[:total_samples]), name
-          end
-        end
-
-        if callees = info[:edges]
-          f.printf "  callees (%d total):\n", info[:total_samples]-info[:samples]
-          callees = callees.map{ |k, weight| [data[:frames][k][:name], weight] }.sort_by{ |k,v| -v }
-          callees.each do |name, weight|
-            f.printf "   % 5d  (% 8s)  %s\n", weight, "%3.1f%%" % (100.0*weight/(info[:total_samples]-info[:samples])), name
-          end
-        end
-
-        f.puts "  code:"
-        source_display(f, file, lines, line-1..maxline)
-      end
     end
 
     # Walk up and down the stack from a given starting point (name).  Loops
